Refactor configuration and dependencies
- Moved ALLOWED_HOSTS to a new config file for better organization.
- Added open-next.config.ts for Cloudflare configuration.
- Updated package.json to include OpenNext.js Cloudflare dependencies and scripts for preview and deployment.
- Removed Vercel analytics and speed insights from layout.
- Adjusted animation properties in LevelPicker component for improved UX.
- Fixed className order in WeaponThumbnail component for consistency.
- Created wrangler.toml for Cloudflare Workers configuration.
