+++ .gitignore
@@ -37,6 +37,9 @@ yarn-error.log*
 # vercel
 .vercel
 
+#open next
+.open-next
+
 # typescript
 *.tsbuildinfo
 next-env.d.ts
