+++ src/app/api/image/route.ts
@@ -4,16 +4,7 @@ import { LRUCache } from "lru-cache";
 import { NextResponse } from "next/server";
 import sharp from "sharp";
 
-// Allowed hosts for security (replace with your external service domains)
-export const ALLOWED_HOSTS = [
-  "homdgcat.wiki",
-  "gi.yatta.moe",
-  "cdn.teyvatarchive.online",
-  "enka.network",
-  "sdk.hoyoverse.com",
-  "fastcdn.hoyoverse.com",
-  "act-webstatic.hoyoverse.com"
-];
+import { ALLOWED_HOSTS } from "~/config/allowedHosts";
 
 // In-memory cache
 const cache = new LRUCache<string, Buffer>({
