+++ src/config/allowedHosts.ts
@@ -1,0 +1,10 @@
+// Allowed hosts for security (replace with your external service domains)
+export const ALLOWED_HOSTS = [
+  "homdgcat.wiki",
+  "gi.yatta.moe",
+  "cdn.teyvatarchive.online",
+  "enka.network",
+  "sdk.hoyoverse.com",
+  "fastcdn.hoyoverse.com",
+  "act-webstatic.hoyoverse.com"
+];
