+++ wrangler.toml
@@ -1,0 +1,7 @@
+  main = ".open-next/worker.js"
+  name = "teyvat-archive-client-next"
+  compatibility_date = "2025-03-25"
+  compatibility_flags = ["nodejs_compat"]
+  [assets]
+  directory = ".open-next/assets"
+  binding = "ASSETS"
